preprocessor: actually run execute --check as part of CI
This got missed from https://cuelang.org/cl/1170108.

As did the fact that --check should cause execute to return early,
regardless of error state. If there is an error we should report it. If
there is not an error we should _not_ proceed to execute the
preprocessor. --check is like a vet check. Indeed it should probably
live under a separate command, but that is for another day.

Preprocessor-No-Write-Cache: true
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cuelang.org/+/1170149
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>

+++ _scripts/runPreprocessor.bash
@@ -28,12 +28,12 @@ GOBIN=$PWD/.gobin go install -trimpath -buildvcs=false ./internal/cmd/preprocess
 
 # Remove preprocessor and hugo target directories. This ensures we don't leave
 # any stale files lying around. Slight hack: only do so if we are running the
-# execute command (but not when running with the --serve flag) . Note this ties
-# this script to _scripts/serve.bash so if making changes to order or args in
-# that script, make changes to the check here.
+# execute command (but not when running with the --serve or --check flags) .
+# Note this ties this script to _scripts/serve.bash so if making changes to
+# order or args in that script, make changes to the check here.
 #
 # TODO: make the prepreprocessor smart enough to do this itself.
-if [[ "${1:-}" == "execute" && "${2:-}" != "--serve" ]]; then
+if [[ "${1:-}" == "execute" && "${2:-}" != "--serve" && "${2:-}" != "--check" ]]; then
 	rm -rf _public hugo/content
 fi
